requirements.txt
blinker
bson
cryptography
domcheck
email_split
email-validator
fastapi>=0.68.0
feedgen
feedgenerator
gcloud
geoip2
google-cloud-datastore
google-cloud-pubsub
gunicorn
itsdangerous
jinja2
lxml
mailjet_rest
openai
PySocks
pydantic>=1.8.0
python-dateutil
python-multipart>=0.0.5
python-whois
requests
simplejson
slowapi
starlette>=0.14.0
typing-extensions>=4.0.0
urllib3
uvicorn>=0.15.0
validate_email
validators
whois
pyyaml
ua-parser
user-agents
httpx
ipaddress
python-jose
validate-email
aiohttp
